requirements.txt
Django==4.2.29
psycopg2-binary==2.9.10
python-decouple==3.8
